Extract tcp and user subcommands
Move tcp and user flags, commands, and run functions to separate files.
