+++ cmd/preflight/main.go
@@ -5,14 +5,10 @@ import (
 	"os"
 	"os/exec"
 	"strings"
-	"time"
 
 	"github.com/spf13/cobra"
 
-	"github.com/vertti/preflight/pkg/output"
 	"github.com/vertti/preflight/pkg/preflightfile"
-	"github.com/vertti/preflight/pkg/tcpcheck"
-	"github.com/vertti/preflight/pkg/usercheck"
 )
 
 // Version is set at build time via ldflags
@@ -53,20 +49,6 @@ var rootCmd = &cobra.Command{
 	Version: Version,
 }
 
-var tcpCmd = &cobra.Command{
-	Use:   "tcp <host:port>",
-	Short: "Check TCP connectivity to a host:port",
-	Args:  cobra.ExactArgs(1),
-	RunE:  runTCPCheck,
-}
-
-var userCmd = &cobra.Command{
-	Use:   "user <username>",
-	Short: "Check that a user exists and meets requirements",
-	Args:  cobra.ExactArgs(1),
-	RunE:  runUserCheck,
-}
-
 var runCmd = &cobra.Command{
 	Use:   "run",
 	Short: "Run checks from a .preflight file",
@@ -74,71 +56,11 @@ var runCmd = &cobra.Command{
 	RunE:  runRun,
 }
 
-var (
-	// tcp flags
-	tcpTimeout time.Duration
-
-	// user flags
-	userUID  string
-	userGID  string
-	userHome string
-
-	// run flags
-	runFile string
-)
+var runFile string
 
 func init() {
-	// tcp subcommand
-	tcpCmd.Flags().DurationVar(&tcpTimeout, "timeout", 5*time.Second, "connection timeout")
-	rootCmd.AddCommand(tcpCmd)
-
-	// user subcommand
-	userCmd.Flags().StringVar(&userUID, "uid", "", "expected user ID")
-	userCmd.Flags().StringVar(&userGID, "gid", "", "expected primary group ID")
-	userCmd.Flags().StringVar(&userHome, "home", "", "expected home directory")
-	rootCmd.AddCommand(userCmd)
-
-	// run subcommand
 	runCmd.Flags().StringVar(&runFile, "file", "", "path to .preflight file (default: search up from current directory)")
 	rootCmd.AddCommand(runCmd)
-}
-
-func runTCPCheck(cmd *cobra.Command, args []string) error {
-	address := args[0]
-
-	c := &tcpcheck.Check{
-		Address: address,
-		Timeout: tcpTimeout,
-		Dialer:  &tcpcheck.RealDialer{},
-	}
-
-	result := c.Run()
-	output.PrintResult(result)
-
-	if !result.OK() {
-		os.Exit(1)
-	}
-	return nil
-}
-
-func runUserCheck(cmd *cobra.Command, args []string) error {
-	username := args[0]
-
-	c := &usercheck.Check{
-		Username: username,
-		UID:      userUID,
-		GID:      userGID,
-		Home:     userHome,
-		Lookup:   &usercheck.RealUserLookup{},
-	}
-
-	result := c.Run()
-	output.PrintResult(result)
-
-	if !result.OK() {
-		os.Exit(1)
-	}
-	return nil
 }
 
 func runRun(cmd *cobra.Command, args []string) error {
